Refactor pooling GTest harness to support 3D and fix verification bug
- Update pooling2d_common.hpp to be template-aware of spatial dimensions (SptDim)
- Refactor pooling3d.cpp and pooling3d_ndhwc.cpp to use the unified harness
- Fix 3D verification bug where 2D dimensions were hardcoded in the common path
