+++ projects/miopen/test/gtest/pooling2d_asymmetric.cpp
@@ -1,16 +1,15 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
 // SPDX-License-Identifier:  MIT
 
-#include <gtest/gtest.h>
-#include <half/half.hpp>
-#include <vector>
-#include "pooling2d_common.hpp"
+#include "pooling_common.hpp"
 
 namespace {
 
-std::vector<pooling2d_gtest::PoolingTestCase> GetPooling2dAsymmetricTestCases()
+using PoolingTestCase = pooling_gtest::PoolingTestCase;
+
+std::vector<PoolingTestCase> GetPooling2dAsymmetricTestCases()
 {
-    static std::vector<pooling2d_gtest::PoolingTestCase> cached_test_cases;
+    static std::vector<PoolingTestCase> cached_test_cases;
     static bool cached = false;
 
     if(cached)
@@ -18,36 +17,30 @@ std::vector<pooling2d_gtest::PoolingTestCase> GetPooling2dAsymmetricTestCases()
         return cached_test_cases;
     }
 
-    std::vector<pooling2d_gtest::PoolingTestCase> test_cases;
+    std::vector<PoolingTestCase> test_cases;
 
     // Dataset 1: Asymmetric configurations
-    // Match ctest's generate_multi_data_limited(..., 9)
-    // For Dataset 1, there is only 1 shape, so no actual limiting happens,
-    // but we use the same generation pattern for consistency.
     std::vector<std::vector<int>> dataset1_inputs = {{1, 4, 4, 4}};
-
-    // Lens: {{2, 2}, {1, 2}, {2, 1}} - asymmetric kernel sizes
     std::vector<std::vector<int>> dataset1_lens = {{2, 2}, {1, 2}, {2, 1}};
-
-    // Strides: {{1, 1}, {2, 1}, {1, 2}, {2, 2}} - asymmetric strides
     std::vector<std::vector<int>> dataset1_strides = {{1, 1}, {2, 1}, {1, 2}, {2, 2}};
 
-    // Pads: controlled by WORKAROUND_ISSUE_1670 (matching original ctest behavior)
 #if WORKAROUND_ISSUE_1670
     std::vector<std::vector<int>> dataset1_pads = {{0, 0}};
 #else
     std::vector<std::vector<int>> dataset1_pads = {{0, 0}, {0, 1}, {1, 0}, {1, 1}};
 #endif
 
-    // Match ctest: Dataset 1 only uses miopenIndexUint32
     std::vector<miopenIndexType_t> dataset1_index_types = {miopenIndexUint32};
     std::vector<miopenPoolingMode_t> modes              = {
                      miopenPoolingMax, miopenPoolingAverage, miopenPoolingAverageInclusive};
     std::vector<int> wsidx_values = {0, 1};
 
+    int num_uint16_case = 0, num_uint32_case = 0, num_uint32_case_imgidx = 0;
+    int num_uint64_case = 0, num_uint64_case_imgidx = 0;
+
     for(const auto& in_shape : dataset1_inputs)
     {
-        pooling2d_gtest::AddTestCasesForInput(in_shape,
+        pooling_gtest::AddTestCasesForInput(in_shape,
                                               dataset1_lens,
                                               dataset1_strides,
                                               dataset1_pads,
@@ -55,25 +48,29 @@ std::vector<pooling2d_gtest::PoolingTestCase> GetPooling2dAsymmetricTestCases()
                                               modes,
                                               wsidx_values,
                                               test_cases,
-                                              true,   // skip_wide_check
-                                              false); // is_wide_dataset
+                                              num_uint16_case,
+                                              num_uint32_case,
+                                              num_uint32_case_imgidx,
+                                              num_uint64_case,
+                                              num_uint64_case_imgidx,
+                                              false,
+                                              false,
+                                              "NCHW");
     }
 
-    // Cache the results
     cached_test_cases = test_cases;
     cached            = true;
-
     return test_cases;
 }
 
 } // anonymous namespace
 
 // Derived classes for Dataset 1 (asymmetric pooling)
-class GPU_AsymPooling2d_FP32 : public pooling2d_gtest::Pooling2dCommon<float>
+class GPU_AsymPooling2d_FP32 : public pooling_gtest::PoolingCommon<float>
 {
 };
 
-class GPU_AsymPooling2d_FP16 : public pooling2d_gtest::Pooling2dCommon<half_float::half>
+class GPU_AsymPooling2d_FP16 : public pooling_gtest::PoolingCommon<half_float::half>
 {
 };
 
@@ -84,9 +81,9 @@ TEST_P(GPU_AsymPooling2d_FP16, HalfTest_pooling2d_asymmetric) { this->RunTest();
 INSTANTIATE_TEST_SUITE_P(Full,
                          GPU_AsymPooling2d_FP32,
                          testing::ValuesIn(GetPooling2dAsymmetricTestCases()),
-                         pooling2d_gtest::GetPoolingTestCaseName);
+                         pooling_gtest::GetPoolingTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(Full,
                          GPU_AsymPooling2d_FP16,
                          testing::ValuesIn(GetPooling2dAsymmetricTestCases()),
-                         pooling2d_gtest::GetPoolingTestCaseName);
+                         pooling_gtest::GetPoolingTestCaseName);
