+++ projects/miopen/test/gtest/pooling2d_asymmetric_nhwc.cpp
@@ -28,15 +28,15 @@
 #include <miopen/env.hpp>
 #include "get_handle.hpp"
 #include "gtest_common.hpp"
-#include "pooling2d_common.hpp"
+#include "pooling_common.hpp"
 
-namespace pooling2d_asymmetric_nhwc {
+namespace pooling2d_wide_nhwc {
 
-using namespace pooling2d_gtest;
+using namespace pooling_gtest;
 
-struct GPU_AsymPooling2d_NHWC_FP32 : public Pooling2dCommon<float> {};
-struct GPU_AsymPooling2d_NHWC_FP16 : public Pooling2dCommon<half_float::half> {};
-struct GPU_AsymPooling2d_NHWC_BFP16 : public Pooling2dCommon<bfloat16> {};
+struct GPU_WidePooling2d_NHWC_FP32 : public PoolingCommon<float> {};
+struct GPU_WidePooling2d_NHWC_FP16 : public PoolingCommon<half_float::half> {};
+struct GPU_WidePooling2d_NHWC_BFP16 : public PoolingCommon<bfloat16> {};
 
 std::vector<PoolingTestCase> GetTestCases(miopenDataType_t prec)
 {
@@ -44,50 +44,44 @@ std::vector<PoolingTestCase> GetTestCases(miopenDataType_t prec)
     std::vector<miopenPoolingMode_t> modes = {miopenPoolingMax, miopenPoolingAverage, miopenPoolingAverageInclusive};
     std::vector<miopenIndexType_t> index_types = {miopenIndexUint8, miopenIndexUint16, miopenIndexUint32, miopenIndexUint64};
 
-    // Dataset 1: Asymmetric configs
-    std::vector<std::vector<int>> in_shapes = {{1, 4, 4, 4}};
-    std::vector<std::vector<int>> lens_list = {{2, 2}, {1, 2}, {2, 1}};
-    std::vector<std::vector<int>> strides_list = {{1, 1}, {2, 1}, {1, 2}, {2, 2}};
+    // Dataset 2: Wide window configs
+    std::vector<std::vector<int>> in_shapes = {{1, 3, 255, 255}, {2, 3, 227, 227}, {1, 7, 127, 127}, {1, 1, 410, 400}};
+    std::vector<std::vector<int>> lens_list = {{35, 35}, {100, 100}, {255, 255}, {410, 400}};
+    std::vector<std::vector<int>> strides_list = {{1, 1}};
     std::vector<std::vector<int>> pads_list = {{0, 0}};
 
     std::vector<PoolingTestCase> test_cases;
+    int num_uint16_case = 0, num_uint32_case = 0, num_uint32_case_imgidx = 0;
+    int num_uint64_case = 0, num_uint64_case_imgidx = 0;
+
     for(const auto& in_shape : in_shapes)
     {
-        AddTestCasesForInput(in_shape,
-                             lens_list,
-                             strides_list,
-                             pads_list,
-                             index_types,
-                             modes,
-                             wsidx_values,
-                             test_cases,
-                             false, // skip_wide_check
-                             false, // is_wide_dataset
-                             "NHWC",
-                             "NHWC");
+        AddTestCasesForInput(in_shape, lens_list, strides_list, pads_list, index_types, modes, wsidx_values, test_cases,
+                             num_uint16_case, num_uint32_case, num_uint32_case_imgidx, num_uint64_case, num_uint64_case_imgidx,
+                             false, true, "NHWC");
     }
     return test_cases;
 }
 
-} // namespace pooling2d_asymmetric_nhwc
+} // namespace pooling2d_wide_nhwc
 
-using namespace pooling2d_asymmetric_nhwc;
+using namespace pooling2d_wide_nhwc;
 
-TEST_P(GPU_AsymPooling2d_NHWC_FP32, FloatTest) { RunTest(); }
-TEST_P(GPU_AsymPooling2d_NHWC_FP16, HalfTest) { RunTest(); }
-TEST_P(GPU_AsymPooling2d_NHWC_BFP16, BFloat16Test) { RunTest(); }
+TEST_P(GPU_WidePooling2d_NHWC_FP32, FloatTest) { RunTest(); }
+TEST_P(GPU_WidePooling2d_NHWC_FP16, HalfTest) { RunTest(); }
+TEST_P(GPU_WidePooling2d_NHWC_BFP16, BFloat16Test) { RunTest(); }
 
 INSTANTIATE_TEST_SUITE_P(Full,
-                         GPU_AsymPooling2d_NHWC_FP32,
+                         GPU_WidePooling2d_NHWC_FP32,
                          testing::ValuesIn(GetTestCases(miopenFloat)),
                          GetPoolingTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(Full,
-                         GPU_AsymPooling2d_NHWC_FP16,
+                         GPU_WidePooling2d_NHWC_FP16,
                          testing::ValuesIn(GetTestCases(miopenHalf)),
                          GetPoolingTestCaseName);
 
 INSTANTIATE_TEST_SUITE_P(Full,
-                         GPU_AsymPooling2d_NHWC_BFP16,
+                         GPU_WidePooling2d_NHWC_BFP16,
                          testing::ValuesIn(GetTestCases(miopenBFloat16)),
                          GetPoolingTestCaseName);
