+++ projects/miopen/test/gtest/pooling2d_wide_nhwc.cpp
@@ -28,15 +28,15 @@
 #include <miopen/env.hpp>
 #include "get_handle.hpp"
 #include "gtest_common.hpp"
-#include "pooling2d_common.hpp"
+#include "pooling_common.hpp"
 
 namespace pooling2d_wide_nhwc {
 
-using namespace pooling2d_gtest;
+using namespace pooling_gtest;
 
-struct GPU_WidePooling2d_NHWC_FP32 : public Pooling2dCommon<float> {};
-struct GPU_WidePooling2d_NHWC_FP16 : public Pooling2dCommon<half_float::half> {};
-struct GPU_WidePooling2d_NHWC_BFP16 : public Pooling2dCommon<bfloat16> {};
+struct GPU_WidePooling2d_NHWC_FP32 : public PoolingCommon<float> {};
+struct GPU_WidePooling2d_NHWC_FP16 : public PoolingCommon<half_float::half> {};
+struct GPU_WidePooling2d_NHWC_BFP16 : public PoolingCommon<bfloat16> {};
 
 std::vector<PoolingTestCase> GetTestCases(miopenDataType_t prec)
 {
@@ -51,20 +51,14 @@ std::vector<PoolingTestCase> GetTestCases(miopenDataType_t prec)
     std::vector<std::vector<int>> pads_list = {{0, 0}};
 
     std::vector<PoolingTestCase> test_cases;
+    int num_uint16_case = 0, num_uint32_case = 0, num_uint32_case_imgidx = 0;
+    int num_uint64_case = 0, num_uint64_case_imgidx = 0;
+
     for(const auto& in_shape : in_shapes)
     {
-        AddTestCasesForInput(in_shape,
-                             lens_list,
-                             strides_list,
-                             pads_list,
-                             index_types,
-                             modes,
-                             wsidx_values,
-                             test_cases,
-                             false, // skip_wide_check
-                             true,  // is_wide_dataset
-                             "NHWC",
-                             "NHWC");
+        AddTestCasesForInput(in_shape, lens_list, strides_list, pads_list, index_types, modes, wsidx_values, test_cases,
+                             num_uint16_case, num_uint32_case, num_uint32_case_imgidx, num_uint64_case, num_uint64_case_imgidx,
+                             false, true, "NHWC");
     }
     return test_cases;
 }
